sandbox/consumer/requirements.txt
pika==1.3.2
msgpack==1.1.1
PyYAML==6.0.2